moved apollo-client to client code
