+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@babel/runtime": "^7.0.0-beta.38",
     "analytics-node": "^2.1.1",
+    "apollo-cache-inmemory": "^1.2.10",
     "apollo-client": "^1.2.2",
     "apollo-engine": "^0.5.4",
     "apollo-errors": "^1.4.0",
