+++ packages/vulcan-lib/lib/client/apollo-client/cache.js
@@ -1,4 +1,4 @@
-import { InMemoryCache } from 'apollo-client-inmemory'
+import { InMemoryCache } from 'apollo-cache-inmemory'
 
 const cache = new InMemoryCache()
 export default cache
